Kind of a lot in this one
I commented out jquery for now. It's a memory hog and we should avoid it
if we can.

The yelp API now works and I've commented it pretty thoroughly. There
are some extra fun javascriptey things going on in there like closures
and callbacks. But it should be pretty straightforward.

Google maps geocoder and the yelp api have to work together kinda
frequently since yelp often returns addresses instead of lat's and
long's.

We should add some kind of loading notification, I'll open an issue
about it and another bug I've seen.